Converting ResetWallet dialog from OnboardingActivity into a separate DialogFragment in order to be able to easily use in other places
